Move fixtures to a file
